[flang]Fix global variables going missing in optimisation
When using `opt` or `clang` to convert llvm-ir produced by `tco`, globals
that are declared in a fortran module, but not actually used by that module,
the optimiser decides to remove the variable when it is marked `linkonce`.

This patch removes the `linkonce` attribute from module global variables,
and updates related tests.