The main idea is that it is possible to query element address/value of a LoweredExpr without having to care about how the LoweredExpr was lowered (lambda for array elemental ops or fir::ExtendedValue).
It also allows to query dynamic extents/type parameters even if the
lowered expr is based on array lambdas not evaluated yet.

Under the hood, a LoweredExpr can be one of two things:
 -> A MaterializedExpr, which is mainly intended to represent
    lowered designator, that is representing an expression whose
    values exist in memory or SSA registers. The core feature that
    it brings is to deal with all the different shade of how this
    values must be addressed.
 -> An ArrayExprAsAFunction, which is a lambda that generates value
    given some indices, and knows about its type/shape/length parameters.

Then, the ability to create loops over a shape and call some lambda
taking indices is moved in their own high level function, outside of
any expression lowering context. It can take and thread array values
if any are provided to it.

All this allows simplifying Expr<T> lowering into a single framework
LoweredExprBuilder that is much simpler since it does not have to deal
with Designator<T> itself and how to address them.

The plan is to allow all ConvertExpr.h genExprXXX to be replaced
by a lowerExpr = genExpr() follow by a few lowerExpr member function
call to place it in a temp, or box it, or get a value, and address...

Assignment-wise, elemental assignment will use a single genAssign
helper, that can take optional loaded array values and where mask.

Forall non elemental procedure call/allocatable assignment will leverage
genAssign in part, but make an explicit ragged temp in lowering (with two
loop nests).

Most of the parts not yet implemented in LoweredExprBuilder should not
require much effort to move nearly directly from the previous ones (
exception is elemental procedure ref that needs to be slightly adapted
to generate the new lambda).

The remaining key risk points here is to get where masks lambdas as well
as elemental user defined assignment nicely, and to see what happens
when fully replacing the previous expression lowering frameworks.