[rebase][OpenMP] Apply D111051 - Add support for SIMD modifier
Some OpenMP local changes related to SIMD were done in
https://github.com/flang-compiler/f18-llvm-project/pull/791
and touched llvm parts outside of flang. These parts were lost
in the rebase/cannot be applied anymore since the files changed too
much.

The upstreaming of these changes is ongoing in:
https://reviews.llvm.org/D111051
Try as much as possible to apply D111051 in a way that works
with fir-dev environment.