[flang] Fix lowering of function calls that return arrays of derived types
When processing the return value of a function call, the lowering code was
incorrectly calling the template function "genarr()" for "ProcedureRef"
arguments rather than "FunctionRef" arguments.  This caused the return type of
the function to be specified as "llvm::None" which caused downstream problems.

The problem was that the version of "genarr" for "FunctionRef" arguments was
declared in such a way that it only was used for arrays whose elements were
intrinsic types.  I fixed this by implementing a more general version of
"genarr" for "FunctionRef" arguments that works for both intrinsic types and
derived types.