[rebase] lib/Frontend dependency fix
FrontendActions.cpp include various mlir/fir headers that themselves
includes auto generated header files from .td.

It is therefore important that FIR and MLIR headers are generated
before starting to build lib/Frontend (otherwise, builds on beefy
machines fail when not finding the auto generated headers).

A previous fix done in:
https://github.com/flang-compiler/f18-llvm-project/commit/9979896045ff884e9df8f0fd41e462a15803e674
Did not add all dependencies and was not working anymore in out-of-tree
builds because LinalgOdsGen is actually no exported in install builds.