[flang] Add a TODO for calls to procedures declared in outer scopes
When an internal procedure calls a dummy procedure which is declared in
an outer scope, we raise an assertion.  This change simply changes that
assertion to a "TODO" so that users will know that this is not yet
implemented rather than an internal error.

An example program can be seen in issue #1186.