[fix 1164] Make symbolIsGlobal the single point
Previously, the code was handling the promotion of local to global
in non re-entrant program outside of symbolIsGlobal. This prevented
the promotion to work on equivalence members, and also caused
latter usage of symbolIsGlobal to be potentially incorrect.

Move all the re-entrant logic in it, and use the symbol scope
to deduce the re-entrant aspects instead of using a boolean that
has to be threaded and is not always accessible after the PFT analysis.

Remove isRecursive since it has no use anymore.

There is a slight change in one of the namelist change because the namelist
code used symbolIsGlobal and previously thought some symbol promoted to globals
were in local memory.