Visit parse tree instead of scopes to identify captured variables
In the context of internal procedure, lowering was relying on scope
information to identify which symbols from a host procedure are used
inside the internal procedures. This relied on symbols with
semantics::HostAssocDetails to be created by the front end for all
such variables used inside internal procedures. This is not always
done, and would be hard to enforce in the frontend/ is not a feature
the front end needs.

Instead, this patch adds a parse tree visitor that calls a call back
on all symbols referenced in a parse-tree procedure node.

This visitor is used for internal procedure with a callback that uses
scopes to determine if a symbol is captured from the host.

This allows fixing issue 990 (statement functions inside internal
procedures hit symbol is symbol is not mapped assert).

This also allows fixing the assert/TODO hit when dummy procedure
symbols from the host are used in an internal procedure. For this
part, on top of detecting the symbol capture, a new CapturedProcedure
class is added in HostAssociations.cpp to defined how this capture
should be done.