Fix namelist lowering in internal procedure
Host namelist used in internal procedures caused lowering to
fail assert "symbol is not mapped to any IR value".
This was because IO lowering in the internal procedure needs to
access the namelist object symbols, and these were not captured since
they are not referred to directly in the internal procedure parse trees.

Fix this by adding namelist object symbols to the host associated
variable list of the host procedure in lowering.