Attempt to fix an issue with writes after no advancing reads.
The issue is that writes after non advancing read overwrite the
beginning of the record (that was read) with blanks.

This is because formatted input is not changing the unit
furtherPositionInRecord, and later write will just think
it should memset the whole beginning with spaces. See this line:
https://github.com/llvm/llvm-project/blob/6d44387e21a1b78eb7263f3610e7f11a42fecd24/flang/runtime/unit.cpp#L278

It is not clear to me if the fix is to not restrict this memset to
certain cases, or if the fix is for input IO to advance the unit
furtherPositionInRecord somehow.

I tried the later, first by doing of max of the current position and
the further one in ConnectionState::HandleAbsolutePosition/RelativPosition.
It lead to more test crashes, so I instead decided to back-up the
positionInRecord inside the furtherPositionInRecord in ExternalIoStatementState<DIR>::EndIoStatement
for non advancing input IO. It fixes my test but not sure it is
correct to do this.