Finish implementation of INQUIRE(IOLENGTH=) (#1234)
The inquire by output list form of the INQUIRE statement calculates
the number of file storage units that would be required to store the
data of an output list in an unformatted file.  Currently, the result
is arbitrary garbage.  This PR makes the individual IO api calls
needed to get the correct result.  Actual output will be off by an
"element size" factor pending an additional fix to the runtime
implementation of the new calls.