Handle pointer and allocatable loop variable
The compiler was failing with "'fir.convert' op invalid type conversion"
when provided with a loop variable that is a pointer or allocatable.

First issue was that createTemp returned the pointer address instead of
the address in the pointer. Use genExprAddr instead.
Rename the function genLoopVariableAddr since it really is its only
usage (it used to be used in symbol lowering when it was part of the
bridge a long time ago).

Second issue was that genType(sym) that returns the pointer type was
used in convert before stores. Use the mlir::Value type of the variable
being stored to instead.