Use dummy character length inside statement function
Like for normal function interfaces with characters arguments, it is possible
that the dummy and actual argument character length differ. The one to
be used inside the statement function is length of the dummy character.

Put the code from SetLength in replaceScalarCharacterLength helper so
that it can be re-used for statement functions.