[flang][driver] Add support for `-S`
Support for `-S` is added in both `flang-new` and `flang-new -fc1`. The
semantics of this flag are consistent with Clang.

The `EmitObjAction` is renamed as `BackendAction`. The new name more
accurately reflects the fact that this action will primarily run the
code-gen/backend pipeline in LLVM. It also makes more sense as an action
implementing both `-emit-obj` and `-S` (originally it was just
`-emit-obj`).

File genneration in `BackendAction::ExecutionAction` is updated to allow
dumping into a pre-defined output stream (to allow e.g. unit testing).
The corresponding code was moved before the definition of the code-gen
pass pipeline. This change is required to make sure that the output
stream is managed more robustly.