Conditional cycle statement rewriting (#1267)
When rewriting an "if (cond) cycle [name]" statement as a structured
branch, an initial NonLabelDoStmt at the head of the evaluation list
supplies information to determine in advance if the cycle statement is
a match or not.  So make a cycle statement a rewrite candidate only if
the corresponding EndDoStmt, when visited, is guaranteed to match.