Fix visibility issue with interdependent module variables
The function instantiateGlobal was wrongfully assuming it would never be
called on module variables being defined (as opposed to used) in the
current scope.

This is wrong since module variable may take each other addresses in
their initializer, causing instantiateGlobal to be called in order to
instantiate the module variable (create the global if needed +
fir.address_of) in the initializer region of the other module variable.

This caused some module symbols to be given local visibility, which
was an issue when they are used in other compilation units.

Remove this assumption and test if the variable requires external
visibility in instantiateGlobal.