Create a new statement context when lowering a nested expression (#1315)
Lowering for an ac-implied-do array constructor places values in an
"outer" temp, where a constructor value may itself be placed in an
"inner" temp.  The mechanism for placing temps in a context to permit
them to be freed in the right scope sometimes passes a context as an
argument, and sometimes accesses it from a scalar or array host
expression lowering class.  Add subscopes to the StatementContext API
to enable correct placement of free calls.

    print*, [(r([7.0]),i=1,3)] ! "inner" temp for calls to r()
    contains
      function r(x)
        real x(:)
        r = x(1)
      end
    end